+++ ChangeLog.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-- FIX : V13 Compatibility Box Module and GETPOST *06/05/2021* - 1.0.4
+## 1.0
+- FIX: v14 compatibility: `f.tva` (now `total_tva`) and `f.total` (now `total_ht`)
+  \+ module descriptor - *13/10/2021* - 1.0.5
+- FIX: V13 Compatibility Box Module and GETPOST *06/05/2021* - 1.0.4
 
-## Version 1.0